% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_report.R
\name{arnaq.create.report}
\alias{arnaq.create.report}
\title{Creates an ARNAQ report from the premade plots.}
\usage{
arnaq.create.report(
  rmarkdown_template = arnaq.report.template,
  qc.name = qc.name,
  genome.file = genome.file,
  flags = arnaq.plot.flags
)
}
\arguments{
\item{rmarkdown_template}{The path to the report template.}

\item{qc.name}{The full name of the report (including any \code{model.name} appended). Any underscores
will be converted to spaces.}

\item{genome.file}{A character object that describes the genome used. Note this is not used
by any data processing, and is used purely so the report includes the genome/version for
completeness. The default uses whatever \verb{\link{arnaq}} found in the \code{resources.yml} file.}

\item{flags}{The vector of flags that defines which sections are included in the final report.
A suitable vector is automatically created during the main \verb{/link\{arnaq\}} run, under the name
\code{arnaq.plot.flags}, which is the default.}
}
\description{
This creates an html report using the previously generated plots and an appropriate report
template. This is called automatically by \verb{\link{arnaq}}, but can be run separately if you have
altered any of the generated plots to suit your needs.
}
\details{
The \code{arnaq.plot.flags} and \code{all_plots} objects must exist.
}
\seealso{
\verb{\link{arnaq}}
}
