% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_directory.R
\name{arnaq.prepare}
\alias{arnaq.prepare}
\title{Create ARNAQ support files}
\usage{
arnaq.prepare(files = c("resources", "template", "biotypes"), ...)
}
\arguments{
\item{files}{A vector naming which support files you want copies of. The default consists of
all available files. Any unrecognised entries will be ignored.}

\item{...}{Further arguments to be passed to file.copy (such as \code{overwrite=TRUE}).}
}
\description{
Creates copies of important ARNAQ files in the working directory.
}
\details{
\code{arnaq_template_1.0.rmd} is a skeleton rmarkdown object used to generate the report. Besides
using the default version, you can edit to to include branding, citations for the specific
platform and tools used, etc..

\code{biotype_conversion_table.txt} is used to map the full set of Ensembl biotypes to a more
manageable set for plotting. The default version should be fine for most applications, but if you
want finer control over how the categories are collapsed you can create your own based on the
default.

\code{resources.yml} is the file used to give project-wide metadata, and an example file is provided.
}
