% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arnaq.R
\name{arnaq_clear}
\alias{arnaq_clear}
\title{Removes ARNAQ objects}
\usage{
arnaq_clear(remove.all = TRUE)
}
\arguments{
\item{remove.all}{If the default of \code{TRUE}, all of ARNAQ's objects will be removed. Otherwise,
\code{count.data}, \code{gene.masks}, \code{sample.mask} and \code{sample.metadata} will be kept, as the objects
needed for downstream analysis.}
}
\description{
Removes objects that ARNAQ has created from the current session. This is useful to force ARNAQ
to load
files from disk if they have been changed, or to 'clean up' the session for downstream analysis.
Any warnings (such as due to the objects not existing) are suppressed.
}
\details{
It is good practice to put \code{arnaq_clear()} at the top of your script so if you rerun from scratch
you can be sure the data is created afresh.
}
