% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_svg.R
\name{export_all_svg}
\alias{export_all_svg}
\title{Exports all the plots created by arnaq in svg format.}
\usage{
export_all_svg(svg.directory = "QC/svg", qc.name = qc.name)
}
\arguments{
\item{svg.directory}{The directory into which to make the export. If it does not exist, it will
be created, if possible. The default is \code{QC/svg}.}

\item{qc.name}{The prefix to use for all the file names. By default, this is the project name -
underscore - model name as set in \code{resources.yml} and the \code{model.name} argument to
\verb{\link{arnaq}}.}
}
\description{
This function is called automatically when \verb{\link{arnaq}} is called with \code{svg=TRUE}. However,
it can be called after adjusting any of the generated plots to ensure the svg export is of the
correct versions of them.
}
\seealso{
\verb{\link{export_svg}}

\verb{\link{arnaq}}
}
