% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_svg.R
\name{export_svg}
\alias{export_svg}
\title{Exports a single plot as svg.}
\usage{
export_svg(the.plot, filename)
}
\arguments{
\item{the.plot}{The plot to export.}

\item{filename}{The filename to export to.}
}
\description{
A convenience function for exporting svg plots, in a consistent format with bulk exports from
\verb{\link{export_all_svg}}.
}
\seealso{
\verb{\link{export_all_svg}}
}
