% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arnaq.R
\name{make_outlier_mask}
\alias{make_outlier_mask}
\title{Create a mask based on sample names}
\usage{
make_outlier_mask(outlier.names)
}
\arguments{
\item{outlier.names}{A vector of character objects with the names of the samples to
\strong{exclude}.}
}
\value{
A logical vector, \code{TRUE} for samples to include.
}
\description{
Returns a logical vector for the samples in the project that do not match any name in the
supplied vector. This is intended for use with the \code{sample.mask} parameter of the
\verb{\link{arnaq}} function. The names are compared to the Display column of \code{samples.txt}.
}
\seealso{
\verb{\link{arnaq}}
}
