% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_clustering.R
\name{plot_all_clustering}
\alias{plot_all_clustering}
\title{Create a list of four plots to show similarity between samples.}
\usage{
plot_all_clustering(data)
}
\arguments{
\item{data}{The table of read counts.}
}
\value{
A list of the four plots.
}
\description{
The plots are: a heatmap based on Euclidean distance, a dendrogram showing hierarchical
clustering based on Euclidean distance, a heatmap based on Spearman's rank correlation of the
Euclidean distances, and a dendrogram showing hierarchical clustering based on that Spearman's
rank correlation.
}
