% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_violin.R
\name{plot_all_metrics_violin}
\alias{plot_all_metrics_violin}
\title{Creates two violin plots of metrics, based on scale.}
\usage{
plot_all_metrics_violin(count.metrics, group)
}
\arguments{
\item{count.metrics}{A list with two named entries, \code{ByBase} and \code{ByOther}, each containing
a dataframe of metrics with samples as columns.}

\item{group}{The group to split on.}
}
\value{
A list containing two plots.
}
\description{
This assumes that the \code{count.metrics} argument is a list with 2 entries: one named \code{ByBase} for
count-related metrics, and one named \code{ByOther} for proportional metrics. This is in order to
keep the latter readable without putting multiple scales on a single plot, which would be
difficult to do well in an automated fashion.
}
\details{
The violin plots are also separated by a group. \verb{/link\{arnaq\}} uses the first entry in its
\code{treat.group} argument to produce these, run this function separately and regenerate the output
if you want it split by another grouping.
}
\seealso{
\verb{\link{arnaq}}

\verb{\link{arnaq.create.report}}
}
