% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pca.R
\name{plot_all_PCA_groups}
\alias{plot_all_PCA_groups}
\title{Creates a list of lists of PCA plots (coloured by group membership, for provided groups) up to a
certain depth.}
\usage{
plot_all_PCA_groups(PC.norm, PC.vars, sample.metadata, cols, depth = 2)
}
\arguments{
\item{PC.norm}{Normalised PC matrix.}

\item{PC.vars}{Vector of variability explained by the PCs.}

\item{sample.metadata}{The metadata with group membership for the samples.}

\item{cols}{A vector naming which of the columns in \code{sample.metadata} should be used to create PCA plots.}

\item{depth}{The depth up to which to generate PCA plots.}
}
\value{
A list of PCA plots up to the requested depth.
}
\description{
The top-level list will contain one entry per group examined. The sublists will contains PCA
plots of component 1 against each of the other components up to the
supplied depth. Each sample is coloured by group membership of the group for that plot.
}
\seealso{
\verb{\link{plot_PCA_group}}

\verb{\link{plot_PCA_group_set}}

\verb{\link{plot_PCA_ids_set}}
}
