% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_scatter.R
\name{plot_all_scatter_pairs}
\alias{plot_all_scatter_pairs}
\title{Creates a list of scatter plot of reads for two samples.}
\usage{
plot_all_scatter_pairs(
  norm.dat,
  dat.for.filter = norm.dat,
  scatter.pairs,
  log = TRUE,
  filter.threshold = 3
)
}
\arguments{
\item{norm.dat}{The dataframe containing the reads to base the scatterplot on.}

\item{dat.for.filter}{A similar dataframe containing the data to use for filtering. By default
the same as \code{norm.dat}.}

\item{scatter.pairs}{A list of length-2 vectors, each entry in the list giving the names of two
samples to create a scatter plot for.}

\item{log}{If \code{TRUE}, use log2 scales for the axes.}

\item{filter.threshold}{The minimum value to filter on in \code{dat.for.filter}. The default of 3
prevents numerical issues when preparing the plot.}
}
\value{
a scatterplot.
}
\description{
The scatter plots compare the reads for genes across two samples, preferably using normalised
data. A filter is applied for minimum read depth for a gene, optionally using a different
dat matrix (e.g. unnormalised).
}
\seealso{
\verb{\link{plot_all_scatter_pairs}}
}
