% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_barplots.R
\name{plot_counts_barplot}
\alias{plot_counts_barplot}
\title{Create read count barplots}
\usage{
plot_counts_barplot(
  counts,
  title = "Read assignment by category",
  max.samples.per.page = 40,
  proportional = FALSE
)
}
\arguments{
\item{counts}{The read count table, samples as columns.}

\item{max.samples.per.page}{The maximum number of bars a single plot should have. When this is
exceeded, there will be multiple plots produced.}

\item{proportional}{If \code{TRUE}, the plots will show the proportion of each assignment category
for that sample instead of the absolute number.}
}
\value{
A list of plots.
}
\description{
From a read assignment table, produces a list of plots for the samples in the table.
}
