% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_complexity.R
\name{plot_detected_curves}
\alias{plot_detected_curves}
\title{Create a plot showing detected genes}
\usage{
plot_detected_curves(
  masked.cpm.data,
  highlight = TRUE,
  threshold = 8,
  pivot = 1
)
}
\arguments{
\item{masked.cpm.data}{The matrix of cpm data, masked to remove unwanted samples, created
automatically as part of the
main ARNAQ workflow.}

\item{highlight}{When \code{TRUE}, highlighting will be performed. When \code{FALSE}, all samples will be
drawn in colour, and will appear in the legend.}

\item{threshold}{The number of 'worst' samples to highlight, out of the total.}

\item{pivot}{What count to use when determining the worst samples.}
}
\value{
A plot showing detected genes at different count levels, for each sample, with
the least complex samples highlighted.
}
\description{
The plot shows how many genes are detected at certain count levels, to see how number of genes
drops off with higher counts.
}
\details{
To improve readability, this function can select the least complex samples to highlight, and
only show those in the legend. Non-highlighted samples will still have their curves drawn, just
in grey behind the highlighted samples.

In order to determine the 'worst' samples, a representative count is used.
}
