% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ERCC.R
\name{plot_ERCC_errors}
\alias{plot_ERCC_errors}
\title{Creates a plot of errors in fold-change estimates for ERCC spike-ins}
\usage{
plot_ERCC_errors(
  mix1_reads,
  mix2_reads,
  ERCC_names,
  title,
  ERCC.ref.table = ERCC.table
)
}
\arguments{
\item{mix1_reads}{A dataframe of normalised reads for the mix 1 samples.}

\item{mix2_reads}{A dataframe of normalised reads for the mix 2 samples.}

\item{ERCC_names}{A vector of names of the ERCC spike-ins}

\item{title}{The title to give the plot}

\item{ERCC.ref.table}{The table of ERCC spike-in data.}
}
\value{
A plot showing the estimates in fold-change
}
\description{
The fold-change is calculated between two specific samples, and all such pairs are summarised
in this plot, showing the estimate errors against the true values given by the eRCC spike-in
definitions.
}
\seealso{
\verb{\link{plot_all.ERCC.pairs}}
}
