% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ERCC.R
\name{plot_ERCC_fc_scatter}
\alias{plot_ERCC_fc_scatter}
\title{Creates a scatterplot of ERCC spike-in fold-change ratios against the true values}
\usage{
plot_ERCC_fc_scatter(
  read_matrix,
  mix1_column,
  mix2_column,
  ERCC.ref.table = ERCC.table
)
}
\arguments{
\item{read_matrix}{The full normalised count matrix.}

\item{mix1_column}{The name of the sample with ERCC mix 1.}

\item{mix2_column}{The name of the sample with ERCC mix 2.}

\item{ERCC.ref.table}{The table of ERCC spike-in data.}
}
\value{
A scatter plot of the fold-change against exected values.
}
\description{
Takes two samples, one from each mix, and calculates the fold change between them for ERCC
spike-ins, the creates a scatter plot against the true, known fold change values.
}
\seealso{
\verb{\link{plot_all.ERCC.pairs}}
}
