% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ERCC.R
\name{plot_ERCC_observed_scatter}
\alias{plot_ERCC_observed_scatter}
\title{Creates a plot of observed and expected ERCC spike-in concentrations}
\usage{
plot_ERCC_observed_scatter(
  read_matrix,
  sample.metadata,
  mix,
  ERCC.ref.table = ERCC.table
)
}
\arguments{
\item{read_matrix}{The raw count matrix, samples as columns.}

\item{sample.metadata}{The sample metadata.}

\item{mix}{Either 1 or 2, to indicate which ERCC mix was used.}

\item{ERCC.ref.table}{The table of ERCC spike-in data.}
}
\value{
A plot showing observed ERCC spike-in proportions against expected.
}
\description{
This is intended where a single ERCC mix has been used across multiple samples. The observed read
counts are restricted to the ERCC spike-ins and concentration estimates constructed from these
values.
}
\details{
The \code{samples.txt} file (and hence \code{sample.metadata}) should contain ERCC mix data when this function is
used; a column named \code{ERCC} containing the values 1 and 2 to specify which mix.
}
