% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_normalisation.R
\name{plot_meanSdPlot}
\alias{plot_meanSdPlot}
\title{Creates a standard meanSdPlot}
\usage{
plot_meanSdPlot(assay)
}
\arguments{
\item{assay}{Normalised data in a form \code{ggplot2::meanSdPlot} can interpret.}
}
\value{
The plot in question.
}
\description{
A wrapper for ggplot2's equivalent function.
}
