% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_violin.R
\name{plot_metrics_violin}
\alias{plot_metrics_violin}
\title{Creates a violin plot of metrics.}
\usage{
plot_metrics_violin(metric.table, group.data, labels, title, yscale = 1)
}
\arguments{
\item{metric.table}{The table of metrics, each column a sample.}

\item{group.data}{A vector of group membership for the samples.}

\item{labels}{A vector of length 2, giving the axes labels.}

\item{title}{The title for the plot.}

\item{yscale}{What scaling to apply to the data. For example, a \code{yscale} of 100 can be used if
you wnat the y axis units to be percent.}
}
\value{
A violin plot of the metrics.
}
\description{
This creates a violin plot showing metrics (e.g. produced by Picard) for samples, stratified by
group. All metrics in the supplied table will be included in a signle plot.
}
