% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pca.R
\name{plot_PCA_group}
\alias{plot_PCA_group}
\title{Creates a PCA plot (coloured by group).}
\usage{
plot_PCA_group(PC.norm, PC.vars, groups, groupName, components = 1:2)
}
\arguments{
\item{PC.norm}{Normalised PC matrix.}

\item{PC.vars}{Vector of variability explained by the PCs.}

\item{groups}{Vector of group membership for the samples.}

\item{groupName}{The name for this group, displayed in the legend.}

\item{components}{The two components to include.}
}
\value{
A PCA plot.
}
\description{
Each sample is coloured by group membership.
}
\seealso{
\verb{\link{plot_PCA_ids}}

\verb{\link{plot_PCA_group_set}}

\verb{\link{plot_all.PCA.group.set}}
}
