% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pca.R
\name{plot_PCA_ids_set}
\alias{plot_PCA_ids_set}
\title{Creates a list of PCA plots (by sample name) up to a certain depth.}
\usage{
plot_PCA_ids_set(PC.norm, PC.vars, depth = 2)
}
\arguments{
\item{PC.norm}{Normalised PC matrix.}

\item{PC.vars}{Vector of variability explained by the PCs.}

\item{depth}{The depth up to which to generate PCA plots.}
}
\value{
A list of PCA plots up to the requested depth.
}
\description{
The list will contain PCA plots of component 1 against each of the other components up to the
supplied depth. Each sample is represented by their name.
}
\seealso{
\verb{\link{plot_PCA_ids}}

\verb{\link{plot_PCA_group_set}}
}
