% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_biotype.R
\name{plot_pie_biotype}
\alias{plot_pie_biotype}
\title{Create a pie chart showing the distribution of biotypes across observed transcripts.}
\usage{
plot_pie_biotype(gene_counts, main)
}
\arguments{
\item{gene_counts}{Vector of gene counts, named with the biotype.}

\item{main}{The title to give the plot.}
}
\value{
A ggplot2 plot of a pie chart.
}
\description{
This is not weighted by the abundance of reads, and only serves as a breakdown of the biotypes
captured in the data set.
}
